pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/navi_bootstrap"]

[tool.hatch.build.targets.wheel.force-include]
"packs" = "navi_bootstrap/packs"
"schema" = "navi_bootstrap/schema"

[project]
name = "navi-bootstrap"
version = "0.1.1"
description = "Jinja2 rendering engine and template packs for bootstrapping projects"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Nelson Spencer" }
]

dependencies = [
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.20.0",
    "navi-sanitize>=0.1.0",
    "pyyaml>=6.0",
]

[project.scripts]
nboot = "navi_bootstrap.cli:cli"
navi-bootstrap = "navi_bootstrap.cli:cli"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.9.0",
    "mypy>=1.8.0",
    "bandit>=1.7.0",
    "types-jsonschema>=4.20.0",
    "types-PyYAML>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "RUF", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
